+++ src/api/integration/integration_test.go
@@ -80,6 +80,20 @@ func nodeAddress() string {
 	return addr
 }
 
+func nodeUsername() string {
+	return os.Getenv("SKYCOIN_NODE_USERNAME")
+}
+
+func nodePassword() string {
+	return os.Getenv("SKYCOIN_NODE_PASSWORD")
+}
+
+func newClient() *api.Client {
+	c := api.NewClient(nodeAddress())
+	c.SetAuth(nodeUsername(), nodePassword())
+	return c
+}
+
 func mode(t *testing.T) string {
 	mode := os.Getenv("SKYCOIN_INTEGRATION_TEST_MODE")
 	switch mode {
@@ -252,7 +266,7 @@ func TestStableCoinSupply(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cs, err := c.CoinSupply()
 	require.NoError(t, err)
@@ -266,7 +280,7 @@ func TestLiveCoinSupply(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cs, err := c.CoinSupply()
 	require.NoError(t, err)
@@ -285,7 +299,7 @@ func TestVersion(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	v, err := c.Version()
 	require.NoError(t, err)
@@ -298,7 +312,7 @@ func TestVerifyAddress(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cases := []struct {
 		name    string
@@ -350,7 +364,7 @@ func TestStableVerifyTransaction(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	badSigStr := "71f2c01516fe696328e79bcf464eb0db374b63d494f7a307d1e77114f18581d7a81eed5275a9e04a336292dd2fd16977d9bef2a54ea3161d0876603d00c53bc9dd"
 	badSigBytes, err := hex.DecodeString(badSigStr)
@@ -439,7 +453,7 @@ func TestStableNoUnconfirmedOutputs(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cases := []struct {
 		name    string
@@ -515,7 +529,7 @@ func TestStableOutputs(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cases := []struct {
 		name    string
@@ -594,7 +608,7 @@ func TestLiveOutputs(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	// Request all outputs and check that HeadOutputs is not empty
 	// OutgoingOutputs and IncomingOutputs are variable and could be empty
@@ -651,7 +665,7 @@ func TestLiveBlock(t *testing.T) {
 	testKnownBlocks(t)
 
 	// Check the knownBadBlockSeqs
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, seq := range knownBadBlockSeqs {
 		b, err := c.BlockBySeq(seq)
 		require.NoError(t, err)
@@ -660,7 +674,7 @@ func TestLiveBlock(t *testing.T) {
 }
 
 func testKnownBlocks(t *testing.T) {
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cases := []struct {
 		name    string
@@ -776,7 +790,7 @@ func TestLiveBlockVerbose(t *testing.T) {
 	testKnownBlocksVerbose(t)
 
 	// Check the knownBadBlockSeqs
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, seq := range knownBadBlockSeqs {
 		b, err := c.BlockBySeqVerbose(seq)
 		require.NoError(t, err)
@@ -786,7 +800,7 @@ func TestLiveBlockVerbose(t *testing.T) {
 }
 
 func testKnownBlocksVerbose(t *testing.T) {
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cases := []struct {
 		name    string
@@ -931,7 +945,7 @@ func TestStableBlockchainMetadata(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	metadata, err := c.BlockchainMetadata()
 	require.NoError(t, err)
@@ -951,7 +965,7 @@ func TestLiveBlockchainMetadata(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	metadata, err := c.BlockchainMetadata()
 	require.NoError(t, err)
@@ -964,7 +978,7 @@ func TestStableBlockchainProgress(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	progress, err := c.BlockchainProgress()
 	require.NoError(t, err)
@@ -978,7 +992,7 @@ func TestLiveBlockchainProgress(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	progress, err := c.BlockchainProgress()
 	require.NoError(t, err)
@@ -999,7 +1013,7 @@ func TestStableBalance(t *testing.T) {
 		addrs  []string
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cases := []balanceTestCase{
 		{
@@ -1048,7 +1062,7 @@ func TestLiveBalance(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	// Genesis address check, should not have a balance
 	b, err := c.Balance([]string{"2jBbGxZRGoQG1mqhPBnXnLTxK6oxsTf8os6"})
@@ -1084,7 +1098,7 @@ func TestStableUxOut(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cases := []struct {
 		name   string
@@ -1117,7 +1131,7 @@ func TestLiveUxOut(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	// A spent uxout should never change
 	ux, err := c.UxOut("fe6762d753d626115c8dd3a053b5fb75d6d419a8d0fb1478c5fffc1fe41c5f20")
@@ -1132,7 +1146,7 @@ func TestLiveUxOut(t *testing.T) {
 }
 
 func scanUxOuts(t *testing.T) {
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	outputs, err := c.Outputs()
 	require.NoError(t, err)
@@ -1166,7 +1180,7 @@ func TestStableAddressUxOuts(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cases := []struct {
 		name    string
@@ -1210,7 +1224,7 @@ func TestLiveAddressUxOuts(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cases := []struct {
 		name         string
@@ -1259,7 +1273,7 @@ func TestStableBlocks(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	progress, err := c.BlockchainProgress()
 	require.NoError(t, err)
@@ -1331,7 +1345,7 @@ func TestLiveBlocks(t *testing.T) {
 }
 
 func testBlocks(t *testing.T, start, end uint64) *readable.Blocks {
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	blocks, err := c.Blocks(start, end)
 	require.NoError(t, err)
@@ -1365,7 +1379,7 @@ func TestStableBlocksVerbose(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	progress, err := c.BlockchainProgress()
 	require.NoError(t, err)
@@ -1444,7 +1458,7 @@ func TestLiveBlocksVerbose(t *testing.T) {
 }
 
 func testBlocksVerbose(t *testing.T, start, end uint64) *readable.BlocksVerbose {
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	blocks, err := c.BlocksVerbose(start, end)
 	require.NoError(t, err)
@@ -1480,7 +1494,7 @@ func TestStableLastBlocks(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	blocks, err := c.LastBlocks(1)
 	require.NoError(t, err)
@@ -1510,7 +1524,7 @@ func TestLiveLastBlocks(t *testing.T) {
 	if !doLive(t) {
 		return
 	}
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	var prevBlock *readable.Block
 	blocks, err := c.LastBlocks(10)
 	require.NoError(t, err)
@@ -1534,7 +1548,7 @@ func TestStableLastBlocksVerbose(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	blocks, err := c.LastBlocksVerbose(1)
 	require.NoError(t, err)
@@ -1567,7 +1581,7 @@ func TestLiveLastBlocksVerbose(t *testing.T) {
 	if !doLive(t) {
 		return
 	}
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	blocks, err := c.LastBlocksVerbose(10)
 	require.NoError(t, err)
@@ -1595,7 +1609,7 @@ func TestStableNetworkConnections(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	connections, err := c.NetworkConnections()
 	require.NoError(t, err)
 	require.Empty(t, connections.Connections)
@@ -1610,7 +1624,7 @@ func TestLiveNetworkConnections(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	connections, err := c.NetworkConnections()
 	require.NoError(t, err)
 	require.NotEmpty(t, connections.Connections)
@@ -1635,7 +1649,7 @@ func TestNetworkDefaultConnections(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	connections, err := c.NetworkDefaultConnections()
 	require.NoError(t, err)
 	require.NotEmpty(t, connections)
@@ -1650,7 +1664,7 @@ func TestNetworkTrustedConnections(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	connections, err := c.NetworkTrustedConnections()
 	require.NoError(t, err)
 	require.NotEmpty(t, connections)
@@ -1665,7 +1679,7 @@ func TestStableNetworkExchangeableConnections(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	connections, err := c.NetworkExchangeableConnections()
 	require.NoError(t, err)
 
@@ -1678,7 +1692,7 @@ func TestLiveNetworkExchangeableConnections(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	_, err := c.NetworkExchangeableConnections()
 	require.NoError(t, err)
 }
@@ -1721,7 +1735,7 @@ func TestLiveTransaction(t *testing.T) {
 		},
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			tx, err := c.Transaction(tc.txID)
@@ -1800,7 +1814,7 @@ func TestStableTransaction(t *testing.T) {
 		})
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			tx, err := c.Transaction(tc.txID)
@@ -1847,7 +1861,7 @@ func TestLiveTransactionVerbose(t *testing.T) {
 		},
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			tx, err := c.TransactionVerbose(tc.txID)
@@ -1926,7 +1940,7 @@ func TestStableTransactionVerbose(t *testing.T) {
 		})
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			tx, err := c.TransactionVerbose(tc.txID)
@@ -1973,7 +1987,7 @@ func TestLiveTransactionEncoded(t *testing.T) {
 		},
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			testTransactionEncoded(t, c, tc, false)
@@ -2042,7 +2056,7 @@ func TestStableTransactionEncoded(t *testing.T) {
 		})
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			testTransactionEncoded(t, c, tc, true)
@@ -2096,7 +2110,7 @@ func TestLiveTransactions(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	addrs := []string{
 		"2kvLEyXwAYvHfJuFCkjnYNRTUfHPyWgVwKt",
 	}
@@ -2192,7 +2206,7 @@ func TestStableTransactions(t *testing.T) {
 		})
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			txnResult, err := c.Transactions(tc.addrs)
@@ -2214,7 +2228,7 @@ func TestLiveConfirmedTransactions(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cTxsSingle, err := c.ConfirmedTransactions([]string{"2kvLEyXwAYvHfJuFCkjnYNRTUfHPyWgVwKt"})
 	require.NoError(t, err)
@@ -2288,7 +2302,7 @@ func TestStableConfirmedTransactions(t *testing.T) {
 		},
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			txnResult, err := c.ConfirmedTransactions(tc.addrs)
@@ -2354,7 +2368,7 @@ func TestStableUnconfirmedTransactions(t *testing.T) {
 		})
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			txnResult, err := c.UnconfirmedTransactions(tc.addrs)
@@ -2376,7 +2390,7 @@ func TestLiveUnconfirmedTransactions(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cTxsSingle, err := c.UnconfirmedTransactions([]string{"2kvLEyXwAYvHfJuFCkjnYNRTUfHPyWgVwKt"})
 	require.NoError(t, err)
@@ -2405,7 +2419,7 @@ func TestLiveTransactionsVerbose(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	addrs := []string{
 		"2kvLEyXwAYvHfJuFCkjnYNRTUfHPyWgVwKt",
 	}
@@ -2478,7 +2492,7 @@ func TestStableTransactionsVerbose(t *testing.T) {
 		})
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			txnResult, err := c.TransactionsVerbose(tc.addrs)
@@ -2500,7 +2514,7 @@ func TestLiveConfirmedTransactionsVerbose(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cTxsSingle, err := c.ConfirmedTransactionsVerbose([]string{"2kvLEyXwAYvHfJuFCkjnYNRTUfHPyWgVwKt"})
 	require.NoError(t, err)
@@ -2574,7 +2588,7 @@ func TestStableConfirmedTransactionsVerbose(t *testing.T) {
 		},
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			txnResult, err := c.ConfirmedTransactionsVerbose(tc.addrs)
@@ -2640,7 +2654,7 @@ func TestStableUnconfirmedTransactionsVerbose(t *testing.T) {
 		})
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			txnResult, err := c.UnconfirmedTransactionsVerbose(tc.addrs)
@@ -2671,7 +2685,7 @@ func TestLiveUnconfirmedTransactionsVerbose(t *testing.T) {
 	if !doLive(t) {
 		return
 	}
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	cTxsSingle, err := c.UnconfirmedTransactionsVerbose([]string{"2kvLEyXwAYvHfJuFCkjnYNRTUfHPyWgVwKt"})
 	require.NoError(t, err)
@@ -2687,7 +2701,7 @@ func TestStableResendUnconfirmedTransactions(t *testing.T) {
 	if !doStable(t) {
 		return
 	}
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	res, err := c.ResendUnconfirmedTransactions()
 	require.NoError(t, err)
 	require.True(t, len(res.Txids) == 0)
@@ -2697,7 +2711,7 @@ func TestLiveResendUnconfirmedTransactions(t *testing.T) {
 	if !doLive(t) {
 		return
 	}
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	_, err := c.ResendUnconfirmedTransactions()
 	require.NoError(t, err)
 }
@@ -2757,7 +2771,7 @@ func TestStableRawTransaction(t *testing.T) {
 		})
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			txnResult, err := c.RawTransaction(tc.txID)
@@ -2806,7 +2820,7 @@ func TestLiveRawTransaction(t *testing.T) {
 		},
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			txnResult, err := c.RawTransaction(tc.txID)
@@ -2850,7 +2864,7 @@ func TestWalletNewSeed(t *testing.T) {
 		},
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			seed, err := c.NewSeed(tc.entropy)
@@ -2926,7 +2940,7 @@ func TestStableAddressTransactions(t *testing.T) {
 		}...)
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			txns, err := c.AddressTransactions(tc.address)
@@ -2972,7 +2986,7 @@ func TestLiveAddressTransactions(t *testing.T) {
 		},
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
@@ -3000,7 +3014,7 @@ func TestStableRichlist(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	richlist, err := c.Richlist(nil)
 	require.NoError(t, err)
@@ -3050,7 +3064,7 @@ func TestLiveRichlist(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	richlist, err := c.Richlist(nil)
 	require.NoError(t, err)
@@ -3072,7 +3086,7 @@ func TestStableAddressCount(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	count, err := c.AddressCount()
 	require.NoError(t, err)
@@ -3085,7 +3099,7 @@ func TestLiveAddressCount(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	count, err := c.AddressCount()
 	require.NoError(t, err)
@@ -3099,7 +3113,7 @@ func TestStableNoUnconfirmedPendingTransactions(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	txns, err := c.PendingTransactions()
 	require.NoError(t, err)
@@ -3111,7 +3125,7 @@ func TestStablePendingTransactions(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	txns, err := c.PendingTransactions()
 	require.NoError(t, err)
@@ -3134,7 +3148,7 @@ func TestLivePendingTransactions(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	_, err := c.PendingTransactions()
 	require.NoError(t, err)
@@ -3145,7 +3159,7 @@ func TestStableNoUnconfirmedPendingTransactionsVerbose(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	txns, err := c.PendingTransactionsVerbose()
 	require.NoError(t, err)
@@ -3157,7 +3171,7 @@ func TestStablePendingTransactionsVerbose(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	txns, err := c.PendingTransactionsVerbose()
 	require.NoError(t, err)
@@ -3180,7 +3194,7 @@ func TestLivePendingTransactionsVerbose(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	_, err := c.PendingTransactionsVerbose()
 	require.NoError(t, err)
@@ -3193,7 +3207,7 @@ func TestLiveWalletSpend(t *testing.T) {
 
 	requireWalletEnv(t)
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	w, totalCoins, _, password := prepareAndCheckWallet(t, c, 2e6, 2)
 
 	tt := []struct {
@@ -3319,7 +3333,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 
 	requireWalletEnv(t)
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	w, totalCoins, totalHours, password := prepareAndCheckWallet(t, c, 2e6, 20)
 
@@ -3940,8 +3954,8 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					},
 				},
 			},
-			err:  "401 Unauthorized - invalid password",
-			code: http.StatusUnauthorized,
+			err:  "400 Bad Request - invalid password",
+			code: http.StatusBadRequest,
 		})
 
 		cases = append(cases, testCase{
@@ -4069,7 +4083,7 @@ func TestLiveWalletCreateTransactionRandom(t *testing.T) {
 
 	requireWalletEnv(t)
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	w, totalCoins, totalHours, password := prepareAndCheckWallet(t, c, 2e6, 20)
 
@@ -4417,7 +4431,7 @@ func TestCreateWallet(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	w, seed, clean := createWallet(t, c, false, "", "")
 	defer clean()
@@ -4466,7 +4480,7 @@ func TestGetWallet(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	// Create a wallet
 	w, _, clean := createWallet(t, c, false, "", "")
@@ -4483,7 +4497,7 @@ func TestGetWallets(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	// Creates 2 new wallets
 	var ws []api.WalletResponse
@@ -4524,7 +4538,7 @@ func TestWalletNewAddress(t *testing.T) {
 	for i := 1; i <= 30; i++ {
 		name := fmt.Sprintf("generate %v addresses", i)
 		t.Run(name, func(t *testing.T) {
-			c := api.NewClient(nodeAddress())
+			c := newClient()
 			var encrypt bool
 			var password string
 			// Test wallet with encryption only when i == 2, so that
@@ -4564,7 +4578,7 @@ func TestStableWalletBalance(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	w, _, clean := createWallet(t, c, false, "", "casino away claim road artist where blossom warrior demise royal still palm")
 	defer clean()
 
@@ -4582,7 +4596,7 @@ func TestLiveWalletBalance(t *testing.T) {
 
 	requireWalletEnv(t)
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	_, walletName, _ := getWalletFromEnv(t, c)
 	bp, err := c.WalletBalance(walletName)
 	require.NoError(t, err)
@@ -4595,7 +4609,7 @@ func TestWalletUpdate(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	w, _, clean := createWallet(t, c, false, "", "")
 	defer clean()
 
@@ -4613,7 +4627,7 @@ func TestStableWalletUnconfirmedTransactions(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	w, _, clean := createWallet(t, c, false, "", "")
 	defer clean()
 
@@ -4631,7 +4645,7 @@ func TestLiveWalletUnconfirmedTransactions(t *testing.T) {
 
 	requireWalletEnv(t)
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	w, _, _, _ := prepareAndCheckWallet(t, c, 1e6, 1)
 	txns, err := c.WalletUnconfirmedTransactions(w.Filename())
 	require.NoError(t, err)
@@ -4652,7 +4666,7 @@ func TestStableWalletUnconfirmedTransactionsVerbose(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	w, _, clean := createWallet(t, c, false, "", "")
 	defer clean()
 
@@ -4670,7 +4684,7 @@ func TestLiveWalletUnconfirmedTransactionsVerbose(t *testing.T) {
 
 	requireWalletEnv(t)
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	w, _, _, _ := prepareAndCheckWallet(t, c, 1e6, 1)
 	txns, err := c.WalletUnconfirmedTransactionsVerbose(w.Filename())
 	require.NoError(t, err)
@@ -4691,7 +4705,7 @@ func TestWalletFolderName(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	folderName, err := c.WalletFolderName()
 	require.NoError(t, err)
 
@@ -4704,7 +4718,7 @@ func TestEncryptWallet(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	// Create a unencrypted wallet
 	w, _, clean := createWallet(t, c, false, "", "")
@@ -4742,13 +4756,13 @@ func TestDecryptWallet(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	w, seed, clean := createWallet(t, c, true, "pwd", "")
 	defer clean()
 
 	// Decrypt wallet with different password, must fail
 	_, err := c.DecryptWallet(w.Meta.Filename, "pwd1")
-	assertResponseError(t, err, http.StatusUnauthorized, "401 Unauthorized - invalid password")
+	assertResponseError(t, err, http.StatusBadRequest, "400 Bad Request - invalid password")
 
 	// Decrypt wallet with no password, must fail
 	_, err = c.DecryptWallet(w.Meta.Filename, "")
@@ -4789,7 +4803,7 @@ func TestRecoverWallet(t *testing.T) {
 
 	// Create an encrypted wallet with some addresses pregenerated,
 	// to make sure recover recovers the same number of addresses
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	wf, err := c.WalletFolderName()
 	require.NoError(t, err)
 
@@ -4855,7 +4869,7 @@ func TestGetWalletSeedDisabledAPI(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	// Create an encrypted wallet
 	w, _, clean := createWallet(t, c, true, "pwd", "")
@@ -4870,7 +4884,7 @@ func TestGetWalletSeedEnabledAPI(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	// Create an encrypted wallet
 	w, seed, clean := createWallet(t, c, true, "pwd", "")
@@ -4890,7 +4904,7 @@ func TestGetWalletSeedEnabledAPI(t *testing.T) {
 
 	// Check with invalid password
 	_, err = c.WalletSeed(w.Meta.Filename, "wrong password")
-	assertResponseError(t, err, http.StatusUnauthorized, "401 Unauthorized - invalid password")
+	assertResponseError(t, err, http.StatusBadRequest, "400 Bad Request - invalid password")
 
 	// Check with missing password
 	_, err = c.WalletSeed(w.Meta.Filename, "")
@@ -5257,7 +5271,7 @@ func TestDisableWalletAPI(t *testing.T) {
 		},
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	for _, tc := range tt {
 		f := func(tc testCase) func(t *testing.T) {
 			return func(t *testing.T) {
@@ -5310,7 +5324,7 @@ func TestStableHealth(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	r, err := c.Health()
 	require.NoError(t, err)
@@ -5338,7 +5352,7 @@ func TestLiveHealth(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 
 	r, err := c.Health()
 	require.NoError(t, err)
@@ -5356,7 +5370,108 @@ func TestDisableGUIAPI(t *testing.T) {
 		return
 	}
 
-	c := api.NewClient(nodeAddress())
+	c := newClient()
 	err := c.Get("/", nil)
 	assertResponseError(t, err, http.StatusNotFound, "404 Not Found")
+}
+
+func TestInvalidAuth(t *testing.T) {
+	if !doLiveOrStable(t) {
+		return
+	}
+
+	user := nodeUsername()
+	pass := nodePassword()
+
+	c := newClient()
+
+	require.Equal(t, user, c.Username)
+	require.Equal(t, pass, c.Password)
+
+	if user == "" && pass == "" {
+		// If neither user nor pass are set, assume the node is run without auth,
+		// and check that providing auth returns a 401 error
+		testInvalidAuthNoAuthConfigured(t)
+	} else {
+		// If either user or pass are set, assume the node is run with auth,
+		// and check that missing or invalid auth returns a 401 error
+		testInvalidAuthAuthConfigured(t)
+	}
+}
+
+func testInvalidAuthAuthConfigured(t *testing.T) {
+	cases := []struct {
+		user string
+		pass string
+	}{
+		{}, // both missing
+		{
+			user: nodeUsername(), // user right, pass missing
+		},
+		{
+			pass: nodePassword(), // pass right, user missing
+		},
+		{
+			user: nodeUsername() + "x", // user wrong, pass missing
+		},
+		{
+			pass: nodePassword() + "x", // pass wrong, user missing
+		},
+		{
+			user: nodeUsername() + "x", // both wrong
+			pass: nodePassword() + "x",
+		},
+		{
+			user: nodeUsername(), // user right, pass wrong
+			pass: nodePassword() + "x",
+		},
+		{
+			user: nodeUsername() + "x", // user wrong, pass right
+			pass: nodePassword(),
+		},
+	}
+
+	for _, tc := range cases {
+		name := fmt.Sprintf("u=%s p=%s", tc.user, tc.pass)
+		t.Run(name, func(t *testing.T) {
+			c := newClient()
+			c.SetAuth(tc.user, tc.pass)
+			require.Equal(t, tc.user, c.Username)
+			require.Equal(t, tc.pass, c.Password)
+
+			_, err := c.Health()
+			assertResponseError(t, err, http.StatusUnauthorized, "401 Unauthorized")
+		})
+	}
+}
+
+func testInvalidAuthNoAuthConfigured(t *testing.T) {
+	cases := []struct {
+		user string
+		pass string
+	}{
+		{
+			user: "foo",
+		},
+		{
+			pass: "bar",
+		},
+		{
+			user: "foo",
+			pass: "bar",
+		},
+	}
+
+	for _, tc := range cases {
+		name := fmt.Sprintf("u=%s p=%s", tc.user, tc.pass)
+		t.Run(name, func(t *testing.T) {
+			c := newClient()
+			c.SetAuth(tc.user, tc.pass)
+			require.Equal(t, tc.user, c.Username)
+			require.Equal(t, tc.pass, c.Password)
+
+			_, err := c.Health()
+			assertResponseError(t, err, http.StatusUnauthorized, "401 Unauthorized")
+		})
+	}
 }